{{cookiecutter.project_slug}}/tests_requirements.txt
coverage==4.4.1
mock==2.0.0
pbr==3.0.1
pluggy==0.4.0
pycodestyle==2.3.1
pytest-cov==2.5.1
pytest-html==1.14.2
pytest-metadata==1.5.0
pytest-travis-fold==1.2.0
tox==2.7.0
virtualenv==15.1.0
